Commit CatBoostClassifier model building
This change adds a Jupyter Notebook (which should be run in colab)
that experiments with different types of models to determine which
should be used to predict customer churn.

It also contains a python file for creating a catboost model
locally for use with the API that will be created in an upcoming
project update.